+++ src/main/groovy/fr/nereide/dom/element/screen/ScreenWidget.groovy
@@ -1,0 +1,10 @@
+package fr.nereide.dom.element.screen
+
+import com.intellij.util.xml.DomElement
+import com.intellij.util.xml.SubTagList
+
+interface ScreenWidget extends DomElement {
+
+    @SubTagList('include-screen')
+    List<IncludeScreen> getIncludeScreens()
+}