+++ src/test/groovy/fr/nereide/test/completion/TestCompletionInXml.groovy
@@ -131,4 +131,10 @@ class TestCompletionInXml extends BaseComplTestCase {
         List<String> expected = ['E1', 'E2', 'E3']
         doTest(expected)
     }
+
+    void testSectionCompletionInFindSimpleScreen() {
+        List<String> expected = ["search-options", "search-results"]
+        doTest(expected)
+    }
+
 }